Change to double

+++ lib/src/cluster.dart
@@ -4,7 +4,7 @@ import 'package:google_maps_flutter_platform_interface/google_maps_flutter_platf
 class Cluster<T extends ClusterItem> {
   final LatLng location;
   final Iterable<T> items;
-  int? zoom;
+  double? zoom;
 
   Cluster(this.items, this.location, {this.zoom});
